Support for 4-connected-neighbor island detection
This commit adds the ability to use the  4-connected neighbor
approach rather than the 8-connected neighbor approach when
finding components in a layer during island evaluation.  This
allows individual pixesls or groups of pixels that are only
touching on the diagonal of a single pixel to be evaluated
independently for their island status rather than being treated
as a single component.

This commit replaces the calls to findContour and drawContour
in LayerManger with a call to connectedComponents, and updates
the supporting logic to work with connectedComponents.  Only
the section of the code dealing specificlly with island
detection has been updated, so areas dealing with resinTraps,
for example, still rely on findContour and drawContour, though,
in theory, connectedComponets would also work for these sections
of the code.

A function for AllowDiagonalBonds has also been added to the
islandDetectionConfig in LayerIssues, but this config item
has not been plumbed up into the GUI as part of this commit.
Changing the default value of AllowDiagonalBonds from false
to true in this code will result in equivilent island detection
to what exsited prior to this commit, albeit via slightly
different logic.