Fix Issue List update on Layer Preview Select
When selecting issues using SHIFT in the layer preview,
the selected issue doesn't update in the issue list until
after shift is released.  This fix forces an immediate
refresh of the issue list so that it updates as you click,
rather than after SHIFT is released.

Also fixed some commented code for for clearing crosshairs
before zoom so that it compiles when uncommented...but code
is still commented out in the commit.