Swap Location of Zoom out to Fit & Autozoom options
After removal of the "Zoom out Large Issues" option in the settings
checkbox, I was just glancing at the settings and thought the
Zoom group would look bettter with the dropdowns for
Zoom In and Zoom Out at the top, and the checkbox
option to enable zoom on single click at the bottom.

Very minor adjustment, so take it or leave it, either way is fine,
but I went ahead and changed it to see how it looked, liked it,
so figured I'd submit.