Fix inconsistencies on pixel edit apply confirm dialog
The same confirmation dialog in the pixel edit mode is used for both
applying operations, and exiting pixel edit mode.

As a result, the text promt feels a little out of context for the apply
case, and this also results in pixel history getting cleard if you
choose "No" on this dialog.  This makes sense if you are exiging pixel
edit mode, but not if you just hit apply and changed your mind.

Added a seperate dialog for the Apply case that only has the options
"Yes/No".  Yes applies the changes, No behaves like Cancel.

Also added a new setting in the settings dialog so that pixel editor
can be configured to exit after each apply operation.  I found I was
applying operations and then going off and doing other things without
realizing I was still in pixel edit mode, so thought this would be
a good option to have.