Customize cursor for Pixel Edit mode
 *(Add)- Support for color cursors via CursorResourceLoader
The built-in Cursor class has a limitation that prevents it from
supporting color or greyscale cursors.  Loading anything other than a 2-bit
cursor will result in an all-black cursor.  Added a CursorResourceLoader
class to work around this issue.
 *(Add) Utilized CursorResourceLoader to load a custom 'pencil' cursor for the
pixel edit mode.
 *(Add) Option to the settings dialog to allow the preview layer
tooltips for select, ROI, and edit pixel mode to be hidden.