Account for crypto padding bug in Chitubox 1.9
In cases where a layer's RLE data is < 0x200 and not a multiple of 0x10, and it gets selected for encryption by Chitubox, it gets stored incorrectly and leads to a corrupt final block. This corrupted block is rarely valid RLE encoded data and will likely cause a parsing exception.

Instead, we look for the closes previous layer that wasn't encrypted (or was able to decode successfully), and clone it to the current layer. For Layer 0 we look forward, for all other layers we look backward.

In cases where all layers are encrypted and affected by the bug, as well as should we for some reason be unable to find a layer to clone, a FileLoadException is thrown.