fix Airmap logic bug / suction cup bug
Adjust airmap logic during 2nd pass to account for contours that were determined to be air-connected during first pass.

Always move resolved resin traps to suction traps, and only conditionally report them as issues if configured to do so. Suction traps are used during a cleanup stage for resin traps, so needs to be populated no matter what, there is no additional processing penalty for this.