Fixed issue of contour duplication
When looking for candidate groups to check intersection  of contours, we used to restrict this to only groups whose last item's layer was 1 above tthe current layer. However, in the event that 2 contours on the same layer intersect with the same group end up in a situation where the "last item" has a layer index of the current layer.

To fix this we loosen the restriction and now scan groups that have a lowest layer index of either current layer index or current layer index + 1