Create new Mat object for resizing mat
During resize, we need to explicitly use a new Mat object and cannot just output the result to aboveLayer. This is because earlier in the code there is a curLayer = aboveLayer, making them both reference the same object, and thus the resize overwites both. This only gets works as it propagates and you end up with aboveLayer, curLayer, belowLayer all referencing the same object. Neighbor checking then never finds any Top or Bottom faces.